requirements.txt
fastapi[standard]~=0.115.2
uvicorn~=0.31.1
PyJWT~=2.9.0
passlib~=1.7.4
pydantic~=2.9.2
bcrypt~=4.0.0
python-multipart~=0.0.12
setuptools~=75.1.0
beanie~=1.27.0
pymongo~=4.9.2